+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+* Make VCardStack's `ìtems` variable visible and remove list modification functions
+* Add VCardNameParameter type for name parameters
+
 ## 1.0.0
 
 * Initial release